lambda/requirements.txt
aws_xray_sdk==2.14.0
boto3==1.34.135
html2text==2024.2.26
#pdf2image==1.17.0
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
amazon-textract-textractor==1.8.2
amazon-textract-textractor[pdf]==1.8.2
pillow
pybars3